Defined TileMapManager to convert the existing tileMap into a 1D array of bytes (multidimensional arrays not serializable by MessagePack). Able to export tilemap and re-import it in sequence. Later will export array to a file (or other network format) and point to/import a map from a file.
Defined a value in TileMapManager referred to as "gridTrimOffset" which is the distance from the play area to set a default tile (currently water). Can now handle a dynamic value (currently hardcoded in initialize).

Modified MasterGrid to accept the byte array and convert it into a 2D grid which defines the terrain.