Scrapping my local dev setup for the backend. Going to be using Bolt to make changes to the live supabase instance and query that. It's going much better.
Modified the maps table to have proper data types for serialization and storage and deserialization (hopefully). I'm again able to send map data to the server to be held in the maps table. trying to resolve issues with converting the data from the server back into the unity defined data structures I created.